cors_dashboard/backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.17.1
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
attrs==25.3.0
bcrypt==3.2.2
bidict==0.23.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.1
cycler==0.12.1
docutils==0.22.2
ecdsa==0.19.1
fastapi==0.117.1
fonttools==4.58.4
frozenlist==1.7.0
gmqtt==0.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
leaflet==0.0.3
Mako==1.3.10
markdown2==2.5.4
MarkupSafe==3.0.2
matplotlib==3.10.3
multidict==6.6.4
nest-asyncio==1.6.0
nicegui==2.24.1
numpy==2.3.1
orjson==3.11.3
packaging==25.0
paho-mqtt==1.6.1
passlib==1.7.4
pillow==11.2.1
propcache==0.3.2
pscript==0.7.7
psutil==7.1.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
PyQt6==6.9.1
PyQt6-Qt6==6.9.2
PyQt6-WebEngine==6.9.0
PyQt6-WebEngine-Qt6==6.9.2
PyQt6_sip==13.10.2
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-jose==3.5.0
python-multipart==0.0.20
python-socketio==5.13.0
PyYAML==6.0.2
requests==2.32.5
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.36.0
vbuild==0.8.2
watchfiles==1.1.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.1
